update on project

+++ app/Http/Controllers/API/ProjectController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class ProjectController extends Controller
 {
@@ -16,22 +17,39 @@ class ProjectController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required|string',
             'description' => 'required|string',
-            'image' => 'required|string',
-            'user_id' => 'required|integer',
+            'start_date' => 'required|date',
+            'end_date' => 'required|date',
         ]);
-        $project = new Project([
-            'name' => $request->name,
-            'description' => $request->description,
-            'image' => $request->image,
-            'user_id' => $request->user_id,
-        ]);
-        $project->save();
-        return response()->json([
-            'message' => 'Successfully created project!',
-        ], 201);
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => $validator->errors()->all(),
+            ]);
+        }
+        try {
+            $project = new Project([
+                'name' => $request->name,
+                'description' => $request->description,
+                'start_date' => $request->start_date,
+                'end_date' => $request->end_date,
+                'live_link' => $request->live_link,
+                'github_link' => $request->github_link,
+                'user_id' => $request->user_id,
+            ]);
+            $project->save();
+            return response()->json([
+                'success' => true,
+                'message' => 'Successfully created project!',
+            ], 201);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => $e->getMessage(),
+            ], 500);
+        }
     }
 
     public function show($id)
@@ -42,25 +60,63 @@ class ProjectController extends Controller
 
     public function update(Request $request, $id)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required|string',
             'description' => 'required|string',
-            'image' => 'required|string',
-            'user_id' => 'required|integer',
+            'start_date' => 'required|date',
+            'end_date' => 'required|date',
         ]);
-        $project = Project::find($id);
-        $project->update($request->all());
-        return response()->json([
-            'message' => 'Successfully updated project!',
-        ], 201);
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => $validator->errors()->all(),
+            ]);
+        }
+        try {
+            $project = Project::find($id);
+            $project->name = $request->name;
+            $project->description = $request->description;
+            $project->start_date = $request->start_date;
+            $project->end_date = $request->end_date;
+            $project->live_link = $request->live_link;
+            $project->github_link = $request->github_link;
+            $project->user_id = auth()->user()->id;
+            $project->save();
+            return response()->json([
+                'success' => true,
+                'message' => 'Successfully updated project!',
+            ], 201);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => $e->getMessage(),
+            ], 500);
+        }
     }
+
 
     public function destroy($id)
     {
         $project = Project::find($id);
-        $project->delete();
-        return response()->json([
-            'message' => 'Successfully deleted project!',
-        ], 201);
+        if (!$project)
+            return response()->json(
+                [
+                    'success' => false,
+                    'message' => 'Project not found!',
+                ],
+                404
+            );
+        try {
+            $project->delete();
+            return response()->json([
+                'success' => true,
+                'message' => 'Successfully deleted project!',
+            ], 201);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => $e->getMessage(),
+            ], 500);
+        }
     }
 }
